// projects/angular-mentions/src/lib/models/coordinate.ts
/**
 * Coordinates for absolute positioning of elements (absolute to top left).
 */
export interface Coordinate {

  /** The number of pixels in regards to the absolute top position. */
  top: number;

  /** The number of pixels in regards to the absolute left position. */
  left: number;

  /** The height in pixels of ??? */
  height?: number;

}